chore(test): refactor v1 bdd steps
In preparation for the implementatino of the V2+ specs, refactor the V1
compatibility suite to share steps
